# docs/swagger/swagger.yaml
basePath: /api
definitions:
...
  realworld-endpoints_internal_dto.UserDTO:
    properties:
      bio:
        type: string
      email:
        type: string
      image:
        type: string
      token:
        type: string
      username:
        type: string
    type: object
  realworld-endpoints_internal_dto.UserLoginParams:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  realworld-endpoints_internal_dto.UserLoginRequest:
    properties:
      user:
        $ref: '#/definitions/realworld-endpoints_internal_dto.UserLoginParams'
    type: object
  realworld-endpoints_internal_dto.UserRegisterParams:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  realworld-endpoints_internal_dto.UserRegisterRequest:
    properties:
      user:
        $ref: '#/definitions/realworld-endpoints_internal_dto.UserRegisterParams'
    type: object
  realworld-endpoints_internal_dto.UserResponse:
    properties:
      user:
        $ref: '#/definitions/realworld-endpoints_internal_dto.UserDTO'
    type: object
host: localhost:8080
info:
  contact: {}
  description: Backend API implementing the RealWorld spec built with Go, Echo Framework,
    GORM, and PostgreSQL.
  title: RealWorld Conduit API
  version: "1.0"
paths:
...
  /user:
    get:
      description: Get the currently authenticated user profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/realworld-endpoints_internal_dto.UserResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get current user
      tags:
      - Users
  /users:
    get:
      description: Returns a list of all registered users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: List all users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Create a new user account and return JWT token
      parameters:
      - description: User registration payload
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/realworld-endpoints_internal_dto.UserRegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/realworld-endpoints_internal_dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      summary: Register a new user
      tags:
      - Users
  /users/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user and return JWT token
      parameters:
      - description: User login credentials
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/realworld-endpoints_internal_dto.UserLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/realworld-endpoints_internal_dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      summary: User login
      tags:
      - Users
securityDefinitions:
  ApiKeyAuth:
    description: 'JWT Token. Format: "Token {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"